Got the new bear button and it's text to appear underneath it's correct brewery.
