+++ src/components/Dashboard.js
@@ -11,15 +11,14 @@ export default () => (
     <>
         <h2>Welcome to NashBrew Your Brewery Catalog</h2>
         <h3>Remember Your best is Different Everyday, but Your Favorite Beer is the Same!</h3>
+        
         <BeerProvider>
-        <BeerTypeProvider>
-            
+            <BeerTypeProvider>
                 <BreweryProvider>
                     <BreweryList />
                     <BeerList />
                 </BreweryProvider>
-            
-        </BeerTypeProvider>
+            </BeerTypeProvider>
         </BeerProvider>
     </>
